+++ docs/make.jl
@@ -15,6 +15,8 @@ makedocs(;
     ),
     pages=[
         "Home" => "index.md",
+        "Manual" => "manual.md",
+        "Library" => "library.md",
     ],
 )